version 2.3.3-3

+++ PKGBUILD
@@ -2,7 +2,7 @@
 pkgname=meshapp-bin
 fullpkgname=meshapp-MeshApp
 pkgver=2.3.3
-pkgrel=2
+pkgrel=3
 pkgdesc="A full-featured desktop client for managing devices and communication in the Meshtastic mesh network."
 arch=('any')
 url="https://git.privatepractice.app/covox/meshapp"
@@ -17,14 +17,19 @@ source=("${url}/releases/download/v${pkgver}/${pkgname%-bin}_${pkgver}_amd64.deb
 #source=("${url}/releases/download/v${pkgver}/MeshApp-v${pkgver}-x86_64.flatpak")
 sha256sums=('c2e00517d562234273a55565dea8a3a42548b16c826d3deaff04156a5ea244fd')
 
-prepare(){
-    bsdtar xzf data.tar.zst -C "${pkgdir}"
-    mkdir -p "${pkgdir}/usr/share/applications/"
-    mv "${pkgdir}/opt/${pkgname%-bin}/lib/${fullpkgname}.desktop" "${pkgdir}/usr/share/applications/${fullpkgname}.desktop"
-    install -D -m644 "${pkgdir}/opt/${pkgname%-bin}/share/doc/copyright" "${pkgdir}/usr/share/licenses/${pkgname%-bin}/LICENSE"
+prepare() {
+    mkdir -p app
+    bsdtar xzf data.tar.zst -C app
+
+    mkdir -p app/usr/share/applications/
+    if [ -f "app/opt/${pkgname%-bin}/lib/${fullpkgname}.desktop" ]; then
+        mv "app/opt/${pkgname%-bin}/lib/${fullpkgname}.desktop" "app/usr/share/applications/${fullpkgname}.desktop"
+    fi
 }
 
-package(){
-    bsdtar xzf data.tar.zst -C "${pkgdir}"
-    install -d "${pkgdir}/opt/${pkgname%-bin}"
+package() {
+    cp -r app/* "${pkgdir}/"
+    if [ -f "${pkgdir}/opt/${pkgname%-bin}/share/doc/copyright" ]; then
+        install -D -m644 "${pkgdir}/opt/${pkgname%-bin}/share/doc/copyright" "${pkgdir}/usr/share/licenses/${pkgname%-bin}/LICENSE"
+    fi
 }
